refactoring of elasticsearch filter test
